added and updated tests
