+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/AvailabilityStrategyUnitTests.cs
@@ -1,0 +1,70 @@
+﻿namespace Microsoft.Azure.Cosmos.Tests
+{
+    using System;
+    using System.Collections.Generic;
+    using System.IO;
+    using System.Net.Http;
+    using System.Threading.Tasks;
+    using Microsoft.Azure.Cosmos;
+    using Microsoft.Azure.Documents;
+    using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+    /// <summary>
+    /// Tests for <see cref="AvailabilityStrategy"/>
+    /// </summary>
+    [TestClass]
+    public class AvailabilityStrategyUnitTests
+    {
+        [TestMethod]
+        public async Task RequestMessageCloneTests()
+        {
+            RequestMessage httpRequest = new RequestMessage(
+                HttpMethod.Get,
+                new Uri("/dbs/testdb/colls/testcontainer/docs/testId", UriKind.Relative));
+
+            string key = Guid.NewGuid().ToString();
+            Dictionary<string, object> properties = new Dictionary<string, object>()
+            {
+                { key, Guid.NewGuid() }
+            };
+
+            RequestOptions requestOptions = new RequestOptions()
+            {
+                Properties = properties
+            };
+
+            httpRequest.RequestOptions = requestOptions;
+            httpRequest.ResourceType = ResourceType.Document;
+            httpRequest.OperationType = OperationType.Read;
+            httpRequest.Headers.CorrelatedActivityId = Guid.NewGuid().ToString();
+            httpRequest.PartitionKeyRangeId = new PartitionKeyRangeIdentity("0", "1");
+            httpRequest.UseGatewayMode = true;
+            httpRequest.ContainerId = "testcontainer";
+            httpRequest.DatabaseId = "testdb";
+            httpRequest.Content = Stream.Null;
+
+            using (CloneableStream clonedBody = await StreamExtension.AsClonableStreamAsync(httpRequest.Content))
+            {
+                RequestMessage clone = httpRequest.Clone(httpRequest.Trace, clonedBody);
+
+                Assert.AreEqual(httpRequest.RequestOptions.Properties, clone.RequestOptions.Properties);
+                Assert.AreEqual(httpRequest.ResourceType, clone.ResourceType);
+                Assert.AreEqual(httpRequest.OperationType, clone.OperationType);
+                Assert.AreEqual(httpRequest.Headers.CorrelatedActivityId, clone.Headers.CorrelatedActivityId);
+                Assert.AreEqual(httpRequest.PartitionKeyRangeId, clone.PartitionKeyRangeId);
+                Assert.AreEqual(httpRequest.UseGatewayMode, clone.UseGatewayMode);
+                Assert.AreEqual(httpRequest.ContainerId, clone.ContainerId);
+                Assert.AreEqual(httpRequest.DatabaseId, clone.DatabaseId);
+            }
+        }
+
+        [TestMethod]
+        public void HedgedRegionToString()
+        {
+            (string region, Uri regionEndpoint) = ("region", new Uri("https://regionEndpoint-region.documents.azure.com"));
+            string hedgedRegionToString = CrossRegionParallelHedgingAvailabilityStrategy.HedgedRegionToString((region, regionEndpoint));
+
+            Assert.AreEqual("region: https://regionendpoint-region.documents.azure.com/", hedgedRegionToString);
+        }
+    }
+}
